requirements.txt
aiogram~=3.5.0
python-dotenv~=1.0.1
Pyrogram~=2.0.106
pytz~=2024.1